requirements.txt
opencv-contrib-python
numpy
pyttsx3
flask
mysql-connector-python
